refactor(cli): address PR review — dedup component, split ChatApp
- Drop AssistantStreamingText (byte-identical to AssistantMessage); reuse
  AssistantMessage in PendingTurn.
- Split ChatApp (557→281 lines), separating logic from UI:
  - useTranscript (transcript.ts): transcript state + id allocation
  - useChatEvents: agent-event → transcript reducer + reasoning/error bookkeeping
  - useSlashCommands: slash-command parsing/dispatch
