+++ packages/cli/src/chat/ChatApp.tsx
@@ -2,7 +2,7 @@ import type { ModelProvider, ProviderTypes } from '@mokei/model-provider'
 import type { Session } from '@mokei/session'
 import { AgentSession } from '@mokei/session'
 import { Box, Static, Text, useApp, useInput } from 'ink'
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
+import { useCallback, useEffect, useRef, useState } from 'react'
 
 import { AssistantMessage } from './components/AssistantMessage.js'
 import { ConfirmCard } from './components/ConfirmCard.js'
@@ -11,37 +11,16 @@ import { HelpCard } from './components/HelpCard.js'
 import { IconLine } from './components/IconLine.js'
 import { ModelSelectCard } from './components/ModelSelectCard.js'
 import { PendingTurn } from './components/PendingTurn.js'
-import { SystemNotice, type SystemNoticeVariant } from './components/SystemNotice.js'
+import { SystemNotice } from './components/SystemNotice.js'
 import { ToolResultCard } from './components/ToolResultCard.js'
 import { ToolSelectCard } from './components/ToolSelectCard.js'
 import { UserMessage } from './components/UserMessage.js'
 import { type AgentSessionLike, useAgentTurn } from './hooks/useAgentTurn.js'
+import { useChatEvents } from './hooks/useChatEvents.js'
 import { useSession } from './hooks/useSession.js'
+import { type ChatModal, useSlashCommands } from './hooks/useSlashCommands.js'
 import { useToolApproval } from './hooks/useToolApproval.js'
-import { parseSlash } from './slash.js'
-
-type TranscriptEntry =
-  | { kind: 'user'; id: number; text: string }
-  | { kind: 'assistant'; id: number; text: string }
-  | {
-      kind: 'tool'
-      id: number
-      name: string
-      result?: string
-      error?: string
-      outcome?: 'error' | 'timeout' | 'cancelled'
-      durationMs?: number
-    }
-  | { kind: 'notice'; id: number; variant: SystemNoticeVariant; text: string }
-  | { kind: 'reasoning'; id: number; text: string }
-
-// Distributive omit preserves the discriminated union (plain `Omit<TranscriptEntry, 'id'>`
-// collapses into an intersection that drops variant-specific fields).
-type TranscriptEntryInput = TranscriptEntry extends infer E
-  ? E extends { id: number }
-    ? Omit<E, 'id'>
-    : never
-  : never
+import { useTranscript } from './transcript.js'
 
 export type ChatAppProps<T extends ProviderTypes> = {
   session: Session<T>
@@ -55,8 +34,7 @@ export function ChatApp<T extends ProviderTypes>(props: ChatAppProps<T>) {
   const { session, provider, providerKey, initialModel, timeout } = props
   const { exit } = useApp()
   const [model, setModel] = useState<string | undefined>(initialModel)
-  const [transcript, setTranscript] = useState<Array<TranscriptEntry>>([])
-  const [modal, setModal] = useState<null | 'model' | 'tools' | 'help'>(null)
+  const [modal, setModal] = useState<ChatModal>(null)
   const [confirmRemove, setConfirmRemove] = useState<string | null>(null)
   const [models, setModels] = useState<Array<{ id: string }>>([])
   const modelsPromiseRef = useRef<Promise<Array<{ id: string }>> | null>(null)
@@ -64,16 +42,8 @@ export function ChatApp<T extends ProviderTypes>(props: ChatAppProps<T>) {
   const quitConfirmRef = useRef(false)
   const quitTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null)
   const [pendingPrompt, setPendingPrompt] = useState<string | null>(null)
-  const [showReasoning, setShowReasoning] = useState(true)
-  const toolStartRef = useRef<Map<string, number>>(new Map())
-  const lastErrorDetailRef = useRef<string | null>(null)
-  // Reasoning accumulated during the current turn, committed to lastReasoningRef
-  // when the turn ends so `/reasoning last` can reprint it.
-  const reasoningBufRef = useRef<string>('')
-  const lastReasoningRef = useRef<string>('')
-  // Mirror of showReasoning readable from the (possibly stale) onEvent closure.
-  const showReasoningRef = useRef(showReasoning)
-  showReasoningRef.current = showReasoning
+
+  const { transcript, pushEntry } = useTranscript()
 
   const loadModels = useCallback(() => {
     if (modelsPromiseRef.current == null) {
@@ -92,19 +62,15 @@ export function ChatApp<T extends ProviderTypes>(props: ChatAppProps<T>) {
     })
   }, [loadModels])
 
-  const nextID = useMemo(() => {
-    let n = 0
-    return () => ++n
-  }, [])
-
-  const pushEntry = useCallback(
-    (entry: TranscriptEntryInput) =>
-      setTranscript((prev) => [...prev, { ...entry, id: nextID() } as TranscriptEntry]),
-    [nextID],
-  )
-
   const { contexts, addContext, removeContext } = useSession(session)
   const { pending, approve, deny, toolApprovalFn } = useToolApproval()
+
+  const getToolCount = useCallback(
+    () => session.getToolsForProvider(provider).length,
+    [session, provider],
+  )
+  const { onEvent, showReasoning, setShowReasoning, getLastReasoning, getLastErrorDetail } =
+    useChatEvents<T>({ pushEntry, timeout, getToolCount })
 
   const createAgent = useCallback((): AgentSessionLike<T> => {
     return new AgentSession<T>({
@@ -116,135 +82,7 @@ export function ChatApp<T extends ProviderTypes>(props: ChatAppProps<T>) {
     })
   }, [session, providerKey, model, toolApprovalFn, timeout])
 
-  const turn = useAgentTurn<T>({
-    createAgent,
-    onEvent: (event) => {
-      switch (event.type) {
-        case 'start':
-          reasoningBufRef.current = ''
-          break
-        case 'reasoning-delta':
-          reasoningBufRef.current += event.reasoning
-          break
-        case 'reasoning-complete':
-          // Persist reasoning into the transcript so it stays visible after the
-          // response text, when reasoning display is enabled.
-          if (showReasoningRef.current && event.reasoning !== '') {
-            pushEntry({ kind: 'reasoning', text: event.reasoning })
-          }
-          break
-        case 'text-complete':
-          if (event.text.length > 0) {
-            pushEntry({ kind: 'assistant', text: event.text })
-          }
-          break
-        case 'tool-call-start':
-          toolStartRef.current.set(event.toolCall.id, event.timestamp)
-          break
-        case 'tool-call-complete': {
-          const content = event.result?.content
-          const text = Array.isArray(content)
-            ? (
-                content.find((c: { type: string }) => c.type === 'text') as
-                  | { type: 'text'; text: string }
-                  | undefined
-              )?.text
-            : undefined
-          const startedAt = toolStartRef.current.get(event.toolCall.id)
-          toolStartRef.current.delete(event.toolCall.id)
-          pushEntry({
-            kind: 'tool',
-            name: event.toolCall.name,
-            result: text ?? '',
-            ...(startedAt != null ? { durationMs: event.timestamp - startedAt } : {}),
-          })
-          break
-        }
-        case 'tool-call-error': {
-          const startedAt = toolStartRef.current.get(event.toolCall.id)
-          toolStartRef.current.delete(event.toolCall.id)
-          const outcome =
-            event.error.name === 'ToolCallTimeoutError'
-              ? 'timeout'
-              : event.error.name === 'ToolCallCancelledError'
-                ? 'cancelled'
-                : 'error'
-          lastErrorDetailRef.current = event.error.stack ?? event.error.message
-          pushEntry({
-            kind: 'tool',
-            name: event.toolCall.name,
-            error: event.error.message,
-            outcome,
-            ...(startedAt != null ? { durationMs: event.timestamp - startedAt } : {}),
-          })
-          if (outcome === 'cancelled') {
-            pushEntry({
-              kind: 'notice',
-              variant: 'warning',
-              text: `tool cancelled: ${event.toolCall.name}`,
-            })
-          }
-          break
-        }
-        case 'tool-call-denied':
-          pushEntry({
-            kind: 'notice',
-            variant: 'warning',
-            text: `tool denied: ${event.toolCall.name}${event.reason ? ` — ${event.reason}` : ''}`,
-          })
-          break
-        case 'error': {
-          const err = event.error
-          lastReasoningRef.current = reasoningBufRef.current
-          lastErrorDetailRef.current = err.stack ?? err.message
-          const cause =
-            err.cause instanceof Error ? ` (cause: ${err.cause.name}: ${err.cause.message})` : ''
-          pushEntry({
-            kind: 'notice',
-            variant: 'error',
-            text: `${err.name}: ${err.message}${cause}`,
-          })
-          break
-        }
-        case 'timeout': {
-          const secs = timeout != null ? Math.round(timeout / 1000) : null
-          pushEntry({
-            kind: 'notice',
-            variant: 'warning',
-            text:
-              secs != null
-                ? `turn timed out after ${secs}s — pass --timeout to adjust`
-                : 'turn timed out — pass --timeout to adjust',
-          })
-          lastReasoningRef.current = reasoningBufRef.current
-          break
-        }
-        case 'max-iterations':
-          lastReasoningRef.current = reasoningBufRef.current
-          pushEntry({
-            kind: 'notice',
-            variant: 'warning',
-            text: 'max iterations reached',
-          })
-          break
-        case 'complete':
-          lastReasoningRef.current = reasoningBufRef.current
-          if (event.result.text === '' && event.result.toolCalls.length === 0) {
-            const toolCount = session.getToolsForProvider(provider).length
-            const hint =
-              toolCount === 0
-                ? ' — no tools are enabled for the model to call (add a context and enable its tools with /tools)'
-                : ` — ${toolCount} tool(s) available, but the model produced no text or tool call`
-            pushEntry({
-              kind: 'notice',
-              variant: 'warning',
-              text: `stream ended with no output (finish: ${event.result.finishReason})${hint}`,
-            })
-          }
-          break
-      }
-    },
-  })
+  const turn = useAgentTurn<T>({ createAgent, onEvent })
 
   useEffect(() => {
     if (pendingPrompt != null && model != null && modal == null) {
@@ -263,137 +101,23 @@ export function ChatApp<T extends ProviderTypes>(props: ChatAppProps<T>) {
     }
   }, [turn.state, pending, deny])
 
-  const handleSubmit = useCallback(
-    async (raw: string) => {
-      const parsed = parseSlash(raw)
-      if (parsed.kind === 'message') {
-        if (parsed.text === '') return
-        if (model == null) {
-          setPendingPrompt(parsed.text)
-          await loadModels()
-          setModal('model')
-          pushEntry({ kind: 'notice', variant: 'info', text: 'select a model to continue' })
-          return
-        }
-        pushEntry({ kind: 'user', text: parsed.text })
-        setPendingPrompt(null)
-        await turn.submit(parsed.text)
-        return
-      }
-
-      const { name, args } = parsed
-      switch (name) {
-        case 'help':
-          setModal('help')
-          break
-        case 'quit':
-        case 'exit':
-          exit()
-          break
-        case 'context': {
-          const [sub, ...rest] = args
-          if (sub == null || sub === 'list') {
-            pushEntry({
-              kind: 'notice',
-              variant: 'info',
-              text: contexts.length === 0 ? 'no contexts' : `contexts: ${contexts.join(', ')}`,
-            })
-          } else if (sub === 'add') {
-            const [key, command, ...cmdArgs] = rest
-            if (!key || !command) {
-              pushEntry({
-                kind: 'notice',
-                variant: 'error',
-                text: 'usage: /context add <key> <cmd> [args...]',
-              })
-              break
-            }
-            try {
-              const tools = await addContext({ key, command, args: cmdArgs })
-              pushEntry({
-                kind: 'notice',
-                variant: 'success',
-                text: `context ${key} added (${tools.length} tool(s) enabled — deselect any below)`,
-              })
-              if (tools.length > 0) {
-                setModal('tools')
-              }
-            } catch (err) {
-              pushEntry({ kind: 'notice', variant: 'error', text: (err as Error).message })
-            }
-          } else if (sub === 'remove') {
-            const [key] = rest
-            if (!key) {
-              pushEntry({ kind: 'notice', variant: 'error', text: 'usage: /context remove <key>' })
-              break
-            }
-            if (!contexts.includes(key)) {
-              pushEntry({ kind: 'notice', variant: 'error', text: `unknown context: ${key}` })
-              break
-            }
-            setConfirmRemove(key)
-          } else {
-            pushEntry({
-              kind: 'notice',
-              variant: 'error',
-              text: `unknown: /context ${sub}`,
-            })
-          }
-          break
-        }
-        case 'model': {
-          const [id] = args
-          const list = await loadModels()
-          if (id != null) {
-            if (list.some((m) => m.id === id)) {
-              setModel(id)
-              pushEntry({ kind: 'notice', variant: 'success', text: `model: ${id}` })
-            } else {
-              pushEntry({ kind: 'notice', variant: 'error', text: `unknown model: ${id}` })
-            }
-          } else {
-            setModal('model')
-          }
-          break
-        }
-        case 'tools':
-          setModal('tools')
-          break
-        case 'details':
-          if (lastErrorDetailRef.current == null) {
-            pushEntry({ kind: 'notice', variant: 'info', text: 'no recent error details' })
-          } else {
-            pushEntry({ kind: 'notice', variant: 'info', text: lastErrorDetailRef.current })
-          }
-          break
-        case 'reasoning': {
-          const [arg] = args
-          if (arg === 'last') {
-            pushEntry({
-              kind: 'notice',
-              variant: 'info',
-              text:
-                lastReasoningRef.current === ''
-                  ? 'no reasoning recorded for the last turn'
-                  : lastReasoningRef.current,
-            })
-            break
-          }
-          const next = arg === 'on' ? true : arg === 'off' ? false : !showReasoning
-          setShowReasoning(next)
-          pushEntry({
-            kind: 'notice',
-            variant: 'info',
-            text: `reasoning display: ${next ? 'on' : 'off'}`,
-          })
-          break
-        }
-        default:
-          pushEntry({ kind: 'notice', variant: 'error', text: `unknown command: /${name}` })
-      }
-    },
-    [addContext, contexts, exit, model, provider, pushEntry, removeContext, showReasoning, turn],
-  )
+  const handleSubmit = useSlashCommands({
+    model,
+    setModel,
+    setModal,
+    setConfirmRemove,
+    setPendingPrompt,
+    loadModels,
+    pushEntry,
+    contexts,
+    addContext,
+    submit: turn.submit,
+    exit,
+    showReasoning,
+    setShowReasoning,
+    getLastReasoning,
+    getLastErrorDetail,
+  })
 
   useInput((input, key) => {
     if (key.ctrl && input === 'c') {
